extract cell creation to common base class
